requirements.txt
Django==2.2.6
django-extensions==2.2.5
django-seed==0.1.9
django-silk==3.0.4
djangorestframework==3.10.3
elasticsearch==5.5.3
Faker==2.0.3
psycopg2-binary==2.8.4
